+++ Shuhari.Framework.Common/DomainModel/StopEnumerationException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Shuhari.Framework.DomainModel
+{
+    /// <summary>
+    /// Used to break nested loop when specified condition are met
+    /// </summary>
+    public sealed class StopEnumerationException : Exception
+    {
+    }
+}